requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv
drf-spectacular
django-cors-headers
djangorestframework-simplejwt
djoser
graphene-django
django-graphql-jwt
pillow
Faker